- moved ChatBubble to common folder
